[Sys] Add condition if frame not found and resize image

+++ src/main/python/app/keyframe/face_keyframe.py
@@ -10,6 +10,7 @@ import matplotlib.pyplot as plt
 class FaceKeyframe: 
     def __init__(self):
         self.face_detector = FaceDetector(detector='mtcnn')
+        self.IMG_SIZE = (60, 60)
 
     def _normalize_keypoints(self, point, x, y):
         return (point[0] - x, point[1] - y)
@@ -17,7 +18,7 @@ class FaceKeyframe:
     def face_keyframe_check(self, filename, lower_threshold, higher_threshold):
         img = cv.imread(filename)
         if img is not None:
-            img = cv.cvtColor(img, cv.COLOR_BGR2RGB)
+            img = cv.resize(cv.cvtColor(img, cv.COLOR_BGR2RGB), self.IMG_SIZE, interpolation=cv.INTER_AREA)
             face_and_landmark = self.face_detector.detect_and_extract_landmark(img)
             if face_and_landmark is not None:
                 x, y, w, h = face_and_landmark['box']
@@ -49,36 +50,37 @@ class FaceKeyframe:
         
     def generate_keyframes_from_frames(self, person_id, input_dir_path, output_dir_path):
         person_input_dir_path = os.path.join(input_dir_path, str(person_id))
-        file_paths = os.listdir(person_input_dir_path)
-        file_paths = [os.path.join(person_input_dir_path, file_path) for file_path in file_paths]
-        counter = 0
-        print("generate keyframe from keyframe_id = ", person_id)
-        MAX = 100
+        if os.path.exists(person_input_dir_path):
+            file_paths = os.listdir(person_input_dir_path)
+            file_paths = [os.path.join(person_input_dir_path, file_path) for file_path in file_paths]
+            counter = 0
+            print("generate keyframe from keyframe_id = ", person_id)
+            MAX = 100
 
-        face_images = []
-        LOWER_THRESHOLD = -5
-        HIGHER_THRESHOLD = 5
-        for file_path in file_paths[:MAX]:
-            face_img = self.face_keyframe_check(file_path, LOWER_THRESHOLD, HIGHER_THRESHOLD)
-            if face_img is not None:
-                face_images.append(face_img)
-        
-        keyframe_output_dir_path = os.path.join(output_dir_path, str(person_id))
-        if not os.path.exists(keyframe_output_dir_path):
-            os.makedirs(keyframe_output_dir_path)
+            face_images = []
+            LOWER_THRESHOLD = -5
+            HIGHER_THRESHOLD = 5
+            for file_path in file_paths[:MAX]:
+                face_img = self.face_keyframe_check(file_path, LOWER_THRESHOLD, HIGHER_THRESHOLD)
+                if face_img is not None:
+                    face_images.append(face_img)
+            
+            keyframe_output_dir_path = os.path.join(output_dir_path, str(person_id))
+            if not os.path.exists(keyframe_output_dir_path):
+                os.makedirs(keyframe_output_dir_path)
 
-        face_images = sorted(face_images, key=lambda x: x[1])
-        increment = 1
-        NUMBER_OF_KEYFRAME = 10
-        if len(face_images) > NUMBER_OF_KEYFRAME:
-            increment = len(face_images) // NUMBER_OF_KEYFRAME
-        
-        counter = 0
-        for face_img in face_images[::increment]:
-            img, angle = face_img
-            keyframe_output_path = os.path.join(keyframe_output_dir_path, str(counter) + '.png')
-            img = cv.cvtColor(img, cv.COLOR_RGB2BGR)
-            cv.imwrite(keyframe_output_path, img)
-            counter += 1
-            if counter == NUMBER_OF_KEYFRAME:
-                break
+            face_images = sorted(face_images, key=lambda x: x[1])
+            increment = 1
+            NUMBER_OF_KEYFRAME = 10
+            if len(face_images) > NUMBER_OF_KEYFRAME:
+                increment = len(face_images) // NUMBER_OF_KEYFRAME
+            
+            counter = 0
+            for face_img in face_images[::increment]:
+                img, angle = face_img
+                keyframe_output_path = os.path.join(keyframe_output_dir_path, str(counter) + '.png')
+                img = cv.cvtColor(img, cv.COLOR_RGB2BGR)
+                cv.imwrite(keyframe_output_path, img)
+                counter += 1
+                if counter == NUMBER_OF_KEYFRAME:
+                    break
